made the hud its own file, added "draw" method
