+++ game_loop.py
@@ -8,6 +8,7 @@ from functions import exit_game, text_objects
 import paused
 import game_over
 import new_round
+from hud import Hud
 
 
 class Player(object):
@@ -48,52 +49,6 @@ class Player(object):
         )
 
 
-class Hud(object):
-    def __init__(self, health, score, ammo):
-        self.health = health
-        self.score = score
-        self.image = pygame.image.load("assets/gun_icon.png").convert_alpha()
-
-    def draw_health(self, health):
-        for i in range(health):
-            img_rect = self.image.get_rect(
-                center=(
-                    G.DISPLAY_WIDTH - (30 * (i + 1)),
-                    G.DISPLAY_HEIGHT * 0.97,
-                )
-            )
-            G.SCREEN.blit(self.image, img_rect)
-
-    def draw_score(self, score):
-        scoreboard_surf, scoreboard_rect = text_objects(
-            "Score: " + str(score),
-            G.SMALL_TEXT,
-            G.WHITE,
-            ((G.DISPLAY_WIDTH * 0.058), (G.DISPLAY_HEIGHT * 0.025)),
-        )
-        G.SCREEN.blit(scoreboard_surf, scoreboard_rect)
-
-    def draw_ammo(self, ammo):
-        if ammo == 0:
-            ammo_status = "Reloading"
-        else:
-            ammo_status = str(ammo)
-        ammo_surf, ammo_rect = text_objects(
-            "Ammo: " + ammo_status,
-            G.SMALL_TEXT,
-            G.WHITE,
-            ((G.DISPLAY_WIDTH * 0.87), (G.DISPLAY_HEIGHT * 0.93)),
-        )
-        G.SCREEN.blit(ammo_surf, ammo_rect)
-
-    def draw_controls(self):
-        control_surf, control_rect = text_objects(
-            "Press 'SPACE' to Fire!",
-            G.SMALL_TEXT,
-            G.WHITE,
-            ((G.DISPLAY_WIDTH * 0.5), (G.DISPLAY_HEIGHT * 0.05)),
-        )
-        G.SCREEN.blit(control_surf, control_rect)
 
 
 def game_loop():
@@ -201,10 +156,7 @@ def game_loop():
         G.SCREEN.fill(G.WHITE)
         G.SCREEN.blit(G.BACKGROUND_1.image, G.BACKGROUND_1.rect)
         score = hud.draw_score(player.score)
-        hud.draw_score(player.score)
-        hud.draw_health(player.health)
-        hud.draw_ammo(player.ammo)
-        hud.draw_controls()
+        hud.draw_hud(player.score, player.ammo, player.health)
 
         # Draw all sprites
         all_sprites_list.draw(G.SCREEN)
